Add an expert Hint explaining running cells issue
Noted by Jana in March and recorded as issue #1

I also managed to get my notebook stuck on these cells for quite a while, because I did not know that I NEEDED to prompt the cell five times for it to finish. Maybe a comment explaining how to interrupt a kernel when people see the hourglass icon for their notebooks?